2. Тесты для идеальный вес [#288363]

+++ src/ru/job4j/calculator/Fit.java
@@ -1,21 +1,21 @@
 package ru.job4j.calculator;
 
 public class Fit {
-    public static short manWeight(short height) {
-        short idealWeight = (short) ((height - 100) * 1.15);
+    public static double manWeight(double height) {
+        double idealWeight = (height - 100) * 1.15;
         return idealWeight;
     }
 
-    public static short womanWeight(short height) {
-        short idealWeight = (short) ((height - 110) * 1.15);
+    public static double womanWeight(double height) {
+        double idealWeight = (height - 110) * 1.15;
         return idealWeight;
     }
 
     public static void main(String[] args) {
-        short manHeight = 178;
-        short womanHeight = 160;
-        short manIdealWeight = Fit.manWeight(manHeight);
-        short womanIdealWeight = Fit.womanWeight(womanHeight);
+        double manHeight = 178;
+        double womanHeight = 160;
+        double manIdealWeight = Fit.manWeight(manHeight);
+        double womanIdealWeight = Fit.womanWeight(womanHeight);
         System.out.println("With the man's height of " + manHeight + "cm his ideal weight is " + manIdealWeight + " kilos");
         System.out.println("With the woman's height of " + womanHeight + "cm her ideal weight is " + womanIdealWeight + " kilos");
     }
